dns/requirements.txt
python_whois
dnspython
elasticsearch
scikit_learn
MySQL-python
geoip2
